offr/api/requirements.txt
fastapi
uvicorn
pandas
pydantic
google-genai
